Change booting sequence to cover AArch64.
Change guest booting steps to support AArch64.
The basic difference of booting a VM between X86 and AArch64 is:
- X86 can setup IOAPIC before creating any VCPU.
- AArch64 have to create VCPU's before creating GIC.

The old procedure is:
- load_kernel
    configure_system
- activate_vcpus
    create vcpu
    start vcpu thread

So we have to seperate "activate_vcpus" into "create_vcpus" and
"activate_vcpus" (to start vcpus only). And setup GIC between
the 2 steps.

The new procedure is:
- load_kernel
- create vcpu's
- (Aarch64) enable GIC
- configure_system
- start vcpu's

Change-Id: Ifc1a14f766e4b4c53da0f1e69956da5b793b9fc4
Signed-off-by: Michael Zhao <michael.zhao@arm.com>