Fixed build errors and warnings on AArch64
This is a preparing commit to build and test CH on AArch64. All building
issues were fixed, but no functionality was introduced.
For X86, the logic of code was not changed at all.
For ARM, the architecture specific part is still empty. And we applied
some tricks to workaround lint warnings. But such code will be replaced
later by other commits with real functionality.

Change-Id: I5e0925549e6ad6cb4385a3e45a58a3f02800cc8a
Signed-off-by: Michael Zhao <michael.zhao@arm.com>