arch: aarch64: Add `acpi_rsdp` to kernel cmdline
To specify the entry of ACPI tables, `acpi_rsdp` parameter was appended
to guest kernel cmdline.

In `--cmdline` parameter of CLH, a user need to set `acpi` to `off`
(default), `on` or `force` to work with `acpi_rsdp`.

Since kernel v5.9, after commit 1583052 was merged in kernel, this hack
became impossible. Now we are testing against a modified kernel branch
"ch-5.10.6-aarch64-acpi-testing" in CLH linux fork.

All of this is only a workaround before EFI ready,

Signed-off-by: Michael Zhao <michael.zhao@arm.com>