aarch64: Change memory layout for UEFI & ACPI
The FDT was loaded at the end of RAM. The address of FDT is changable.
While UEFI (edk2 now) requires fixed address to load FDT and RSDP.
Now the FDT is moved to the beginning of RAM, which is a fixed address.
RSDP is wrote to 2 MiB after FDT, also a fixed address.
Kernel comes 2 MiB after RSDP.

Signed-off-by: Michael Zhao <michael.zhao@arm.com>