[TESTING] aarch64: enable UEFI image load
As the UEFI image on arm64 is not formatted, we can't
load UEFI image by current loader.
when load a image, do as the following step:
	1. load it as a kernel in PE format.
	2. if the first step fails for InvalidImageMagicNumber, load it as
an unformatted UEFI image by uefi_loader.

NOTE:
This commit contains a reference to private linux-loader branch, wich
will be removed when the commit on discussion is upstreamed.

Signed-off-by: Jianyong Wu <jianyong.wu@arm.com>